backend/requirements.txt
Django==5.2.13
djangorestframework==3.17.1
psycopg2-binary==2.9.11
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
